requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
playwright>=1.40.0
python-dotenv>=1.0.0